update outdated IBC channels
channel-93 stopped working, after a long search it seems that channel 151 & 152 are working fine. I used both of them so we remember to revert to one of them if the other fails. There is no guarantee though that they will continue working in the future, we probably have to have our own IBC channels